# elearning/e-smart-learn/app/docs/swagger.yaml
definitions:
  apperror.AppError:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      type:
        $ref: '#/definitions/apperror.ErrorType'
    type: object
  apperror.ErrorType:
    enum:
    - not_found
    - bad_request
    - internal_server
    - unauthorized
    - forbidden
    - duplicate_entry
    - validation_error
    type: string
    x-enum-varnames:
    - NotFound
    - BadRequest
    - InternalServer
    - Unauthorized
    - Forbidden
    - DuplicateEntry
    - ValidationError
  consts.BillingCycle:
    enum:
    - monthly
    - yearly
    type: string
    x-enum-varnames:
    - BillingCycleMonthly
    - BillingCycleYearly
...
  dto.ApiResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/apperror.AppError'
        type: array
      metadata: {}
      path:
        type: string
      process_id:
        type: string
      request: {}
      status:
        $ref: '#/definitions/dto.ResponseStatus'
    type: object
...
  dto.CreatePlanRequest:
    properties:
      access_features:
        items:
          type: string
        type: array
      billing_cycle:
        allOf:
        - $ref: '#/definitions/consts.BillingCycle'
        enum:
        - monthly
        - yearly
      currency:
        maxLength: 10
        minLength: 3
        type: string
      description:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        type: number
    required:
    - billing_cycle
    - name
    - price
    type: object
...
  dto.ResponseStatus:
    properties:
      code:
        type: integer
      type:
        type: string
    type: object
...
  dto.UpdatePlanRequest:
    properties:
      access_features:
        items:
          type: string
        type: array
      billing_cycle:
        allOf:
        - $ref: '#/definitions/consts.BillingCycle'
        enum:
        - monthly
        - yearly
      currency:
        maxLength: 10
        minLength: 3
        type: string
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/subscription-plans:
    get:
      consumes:
      - application/json
      description: Return all plans for admin, including inactive plans
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: List all subscription plans
      tags:
      - plans
    post:
      consumes:
      - application/json
      description: Create a subscription plan and sync to Stripe and Postgres
      parameters:
      - description: Create plan payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create subscription plan
      tags:
      - plans
  /api/v1/admin/subscription-plans/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription plan in Postgres and deactivate linked Stripe
        price/product
      parameters:
      - description: Plan ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete subscription plan
      tags:
      - plans
    put:
      consumes:
      - application/json
      description: Update subscription plan and sync changed fields to Stripe and
        Postgres
      parameters:
      - description: Plan ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Update plan payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update subscription plan
      tags:
      - plans
  /api/v1/admin/subscription-plans/{id}/activate:
    put:
      consumes:
      - application/json
      description: Activate plan for new sales
      parameters:
      - description: Plan ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Activate subscription plan
      tags:
      - plans
  /api/v1/admin/subscription-plans/{id}/deactivate:
    put:
      consumes:
      - application/json
      description: Deactivate plan for new sales and set related active subscriptions
        to cancel at period end
      parameters:
      - description: Plan ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Deactivate subscription plan
      tags:
      - plans
...
  /api/v1/subscription-plans/{id}:
    get:
      consumes:
      - application/json
      description: Get a subscription plan detail by id
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get subscription plan by id
      tags:
      - plans
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"